Split dataframe.py into multiple files
